consumer/requirements.txt
pandas==2.0.3
fastavro==1.8.4
requests==2.31.0
hydra-core==1.3.2
pyyaml==6.0.1
confluent_kafka==2.0.2